Progress with the DSL
Refer to the haddocks for the logic behind it.
The goal is to:

- Have a concise & easy to understand DSL for elements (Smart's
designed elements, like accordions etc.)
- Essentially, represent a canvas, on which things can put; and these
things can have arbitrary types.
- Ensure that the contents of a canvas can be /drawn/ in a certain
method: which can be achieved by ensuring constraints are satisfied
for each element in the canvas "list"

--

There's also an example in the module haddocks explaining how to use
this: and the example indicates the use of two types.

--

The earlier approach, using something similar to MarkupM in Blaze was
not suitable for this.

--

If we couldn't have used this, we'd have to use something like:

``` haskell

do
  toMarkup $ myAccordion
  toMarkup $ myButton

```

Which would mean repeated calls to `toMarkup`; cumbersome to say the
least.

--

@noteed