Add support for Alerts.
The rendered Html is incorrect at the moment, which I'll fix once I
have access to multiple screens and the comparison process can be
somewhat automated.
Currently, given the HTML tags are so deeply nested, it becomes
extremely complex to debug what is incorrect.

I'll go over the rendered HTML fragments once I'm back in BE and
programming from my home office. Doing this on a single screen laptop
is impossible.

@noteed