Avatar as a GADT and with type-level guarantees.
[Minor]
We introduce a simple newtype `Image` that is just a wrapper over
`URI`. This is meant to indicate an image.

[Major]
For avatars, as we realise their use in `Cards`. When avatars are embedded in
cards, they must have additional content (title and sub-title (body)).
With the vanilla definition of an Avatar, we couldn't have guaranteed this.

But with GADTs and KindSignatures, we can guarantee this at the type level.

@noteed, this is a notable commit.