Icon lists.
The rendering of icons is fairly broken: all icons must be included in
a div, and the class varies per icon.

For this I'll need to go through the design documentation in more
detail: is there a nomenclature in place to name classes for icons?

@thusc do you have resources you can point me to for clarifying this?

(Also, apologies for using the previous GH handle for mentioning you
in my commit messages, I'll use this handle from now on.)