Started work on modeling grids.
@thusc, this is going to be very interesting.
From what it looks like; we can guarantee, at the type-level, that a
given grid cannot have:

1. Elements lying outside the grid, or with dimensions (elements
spanning multiple rows/cols) that "overflow" from the grid. (DONE in
the GADT constraint)

2. Elements that overlap each other. (TODO)

This can be used to guarantee at the type-level, that an grid we're
constructing is absolutely proven to be correct.

This is where type-level programming becomes theorem proving in
Haskell!