Added more general cartesian type-level functions.
This completes the grid implementation: and this grid implementation
is generaliseable for other grid like operations.

With this, we can mathematically prove that the grids we're
constructing are /sound/ at the compile time: without resulting in
broken layouts at runtime, or worse, runtime errors.

@thusc I believe this is a good use of type-level programming wherein
proofs are encoded in our types, and unsound "values" are impossible
as guaranteed by the compiler.