Nix changes to build the HTML files statically.
@thusc this is the nix derivation that outputs the HTML files.
We rely on in-haskell logic to generate the directory tree in which to
write the files. And I believe we should still do the same, and modify
the generation logic to make it more flexible.

This should do pretty much the same (in terms of building HTML files).

I've also modified the github deployment file to execute nix-build
instead of a custom build shell script.

I believe this should work, without changes.